Fix comparison of signed and unsigned integers
Use the same type for both sides of the comparison.
Found with -Wsign-compare.

+++ libjuicer/sj-metadata.c
@@ -362,7 +362,7 @@ char *
 sj_metadata_helper_lookup_country_code (const char *code)
 {
   const char *country_name;
-  int len, i;
+  gsize len, i;
   /* Musicbrainz uses some additional codes on top of ISO-3166 so
      treat those as a special case if we don't get a match from the
      iso-codes data */
